==> R/5_Alpha_DIv_hiperammon.Rmd <==
---
title: "5_Alpha_DIv_hiperammon"
output: html_document
date: "2023-06-19"
---

```{r libraries and objects}
library("phyloseq"); packageVersion("phyloseq")
library(readr)
library(data.table)
library(ggpubr)
library(phyloseq)
library(tidyverse)
library("ggplot2"); packageVersion("ggplot2")
ps_0 <- readRDS("~/Documents/Lola/Datos_CIPF_2023/Rats_HE_HiperAmmon_model/outputs/ps_0.Rds")
```
#Alpha div plots groups

Alpha-diversity is calculated on the raw data, here data_otu or data_phylo if you are using phyloseq.
It is important to not use filtered data because many richness estimates are modeled on singletons and doubletons in the occurrence table. So, you need to leave them in the dataset if you want a meaningful estimate.
Moreover, we usually not using normalized data because we want to assess the diversity on the raw data and we are not comparing samples to each other but only assessing diversity within each sample.

```{r Counts}
reads_per_sample <- as.matrix(rowSums(t(as.data.frame(otu_table(ps_0)))))
reads_per_sample 
```


```{r Alpha Diversity models to adjust by batch}
require(lme4)
require("lmtest")
library(tidyverse)
Alpha<-estimate_richness(ps_0, measures = c("Observed", "Chao1","Shannon", "Fisher"))

Alpha%>%
    rownames_to_column()%>%
    dplyr::rename(SampleID = rowname)->tmp1

as_tibble(sample_data(ps_0))->tmp2

tmp1<-inner_join(tmp1, tmp2, by="SampleID")
rownames(tmp1)<- tmp1$SampleID
tmp1$SampleID<- NULL
sdt<- tmp1
rm(tmp1,tmp2)

#Differences between groups: Chao1
full<-lme4::lmer(Chao1 ~ Group + (1|Batch), sdt)
null<-lme4::lmer(Chao1 ~ (1|Batch), sdt)
Chao1<-list(lrtest(null, full))
#Likelihood ratio test

#Model 1: Chao1 ~ (1 | Batch)
#Model 2: Chao1 ~ Group + (1 | Batch)
  #Df  LogLik Df Chisq Pr(>Chisq)    
#1   3 -82.345                        
#2   4 -76.850  1 10.99  0.0009161 ***

#Differences between groups: Shannon
full<-lme4::lmer(Shannon ~ Group + (1|Batch), sdt)
null<-lme4::lmer(Shannon ~ (1|Batch), sdt)
Shannon<-list(lrtest(null, full))
#Likelihood ratio test
#Model 1: Shannon ~ (1 | Batch)
#Model 2: Shannon ~ Group + (1 | Batch)
  #Df  LogLik Df  Chisq Pr(>Chisq)  
#1   3  8.8833                       
#2   4 10.4159  1 3.0653    0.07998 .



#Differences between groups: Fisher
full<-lme4::lmer(Fisher ~ Group + (1|Batch), sdt)
null<-lme4::lmer(Fisher ~ (1|Batch), sdt)
Fisher<-list(lrtest(null, full))
#Likelihood ratio test
#Model 1: Fisher ~ (1 | Batch)
#Model 2: Fisher ~ Group + (1 | Batch)
  #Df  LogLik Df  Chisq Pr(>Chisq)   
##1   3 -49.796                        
#2   4 -45.870  1 7.8507    0.00508 **

extract_test_info <- function(test_objects, test_type) {
  results <- lapply(test_objects, function(test) {
    data.frame(
      Test = test_type,
      LogLik_Null = test$LogLik[1],
      LogLik_Full = test$LogLik[2],
      Df = test$Df[2],
      Chisq = test$Chisq[2],
      PValue = test$`Pr(>Chisq)`[2],
      Sig = ifelse(test$`Pr(>Chisq)`[2] < 0.001, "***",
                           ifelse(test$`Pr(>Chisq)`[2] < 0.01, "**",
                                  ifelse(test$`Pr(>Chisq)`[2] < 0.05, "*", "")))
    )
  })
  return(bind_rows(results))
}

# Extract information and create results data frames
chao1_results <- extract_test_info(Chao1, "Chao1")
shannon_results <- extract_test_info(Shannon, "Shannon")
fisher_results <- extract_test_info(Fisher, "Fisher")
# Adjust p-value
chao1_results$adjusted_p_value <- p.adjust(chao1_results$PValue, method = "BH") 
shannon_results$adjusted_p_value <- p.adjust(shannon_results$PValue, method = "BH") 
fisher_results$adjusted_p_value <- p.adjust(fisher_results$PValue, method = "BH") 
#Add asterisks
#Chao1
chao1_results$Sig_adj <- ifelse(chao1_results$adjusted_p_value < 0.001, "***",
                           ifelse(chao1_results$adjusted_p_value < 0.01, "**",
                                  ifelse(chao1_results$adjusted_p_value < 0.05, "*", "")))
#Shannon
shannon_results$Sig_adj <- ifelse(shannon_results$adjusted_p_value < 0.001, "***",
                           ifelse(shannon_results$adjusted_p_value < 0.01, "**",
                                  ifelse(shannon_results$adjusted_p_value < 0.05, "*", "")))
#Fisher
fisher_results$Sig_adj <- ifelse(fisher_results$adjusted_p_value < 0.001, "***",
                           ifelse(fisher_results$adjusted_p_value < 0.01, "**",
                                  ifelse(fisher_results$adjusted_p_value < 0.05, "*", "")))

#Save
write.csv(chao1_results, "~/Library/CloudStorage/OneDrive-UniversitatdeValencia/Doctorado/Estudios/CIPF_2022/Analisis_resultados_ratas_hiperammon/Rats_HE_HiperAmmon_model/outputs/Tables/Chao1_models.csv")
write.csv(shannon_results, "~/Library/CloudStorage/OneDrive-UniversitatdeValencia/Doctorado/Estudios/CIPF_2022/Analisis_resultados_ratas_hiperammon/Rats_HE_HiperAmmon_model/outputs/Tables/Shannon_models.csv")
write.csv(fisher_results, "~/Library/CloudStorage/OneDrive-UniversitatdeValencia/Doctorado/Estudios/CIPF_2022/Analisis_resultados_ratas_hiperammon/Rats_HE_HiperAmmon_model/outputs/Tables/Fisher_models.csv")
```

###Alpha Diversity plots
```{r AlphaDiversity plots}
metadata<- data.frame(sample_data(ps_0)) 
library(ggsignif)

Chao1<-phyloseq::estimate_richness(ps_0, measures = "Chao1")
metadata<- data.frame(sample_data(ps_0)) 
Chao1$Group<-metadata$Group
pChao1 <-ggplot(Chao1, aes(x=Group, y=Chao1))+
    geom_violin(aes(color = Group), trim = FALSE) +  
    geom_boxplot(width = 0.2) +
    theme_classic() + theme(strip.background = element_blank(), axis.text.x.bottom = element_text(angle = -90), axis.text.x = element_text(size = 16), axis.title = element_text(size = 16), axis.text.y = element_text(size = 14), legend.text = element_text(size = 14), legend.title = element_text(size = 14)) +
    geom_signif(stat="identity",
                data=data.frame(x=c(1), xend=c(2), y=c(360), annotation=c("***")),
                aes(x=x,xend=xend, y=y, yend=y, annotation=annotation)) 
pChao1


Shannon<-phyloseq::estimate_richness(ps_0, measures = "shannon")
metadata<- data.frame(sample_data(ps_0)) 
Shannon$Group<-metadata$Group
pShannon <-ggplot(Shannon, aes(x=Group, y=Shannon))+
    geom_violin(aes(color = Group), trim = FALSE) +  
    geom_boxplot(width = 0.2) +
    theme_classic() + theme(strip.background = element_blank(), axis.text.x.bottom = element_text(angle = -90), axis.text.x = element_text(size = 16), axis.title = element_text(size = 16), axis.text.y = element_text(size = 14), legend.text = element_text(size = 14), legend.title = element_text(size = 14))
pShannon

Fisher<-phyloseq::estimate_richness(ps_0, measures = "fisher")
metadata<- data.frame(sample_data(ps_0)) 
Fisher$Group<-metadata$Group
pFisher <-ggplot(Fisher, aes(x=Group, y=Fisher))+
    geom_violin(aes(color = Group), trim = FALSE) +  
    geom_boxplot(width = 0.2) +
    theme_classic() + theme(strip.background = element_blank(), axis.text.x.bottom = element_text(angle = -90), axis.text.x = element_text(size = 16), axis.title = element_text(size = 16), axis.text.y = element_text(size = 14), legend.text = element_text(size = 14), legend.title = element_text(size = 14))  +
    geom_signif(stat="identity",
                data=data.frame(x=c(1), xend=c(2), y=c(52), annotation=c("**")),
                aes(x=x,xend=xend, y=y, yend=y, annotation=annotation)) 
pFisher

ggarrange(pShannon, pFisher, pChao1, common.legend = TRUE, legend="bottom",nrow = 1)


all_AlphaDiv_boxplots <- ggarrange(pShannon, pChao1, pFisher,common.legend = TRUE, legend="bottom",nrow = 1)

# Modify the legend order in the updated plot
all_AlphaDiv_boxplots <- ggsave(filename = "~/Library/CloudStorage/OneDrive-UniversitatdeValencia/Doctorado/Estudios/CIPF_2022/Analisis_resultados_ratas_hiperammon/Rats_HE_HiperAmmon_model/outputs/Figures/AphaDiversity_models.pdf", width = 7, height = 6)
```
