==> R/6_Batch_eff_remv_and_BetaDiv_hipperammon.Rmd <==
---
title: "6_BetaDiv_hipperammon"
output: html_document
date: "2023-06-20"
---
```{r libraries and objects}
library(phyloseq)
library(dplyr)
library(microbiome)

ps_filtered2 <- readRDS("~/Documents/Lola/Datos_CIPF_2023/Rats_HE_HiperAmmon_model/outputs/ps_filtered2.Rds")
ps_comp <- readRDS("~/Documents/Lola/Datos_CIPF_2023/Rats_HE_HiperAmmon_model/outputs/ps_comp.Rds")
ps_clr <- readRDS("~/Documents/Lola/Datos_CIPF_2023/Rats_HE_HiperAmmon_model/outputs/ps_clr.Rds")

rawDada2_ASVs<-as.data.frame(otu_table(ps_filtered2))
metadata_hiper<-as.data.frame(sample_data(ps_filtered2))
```

#PCoAs before batch effect removal
###PCoA relative abundance: Group and batch
```{r PCoA representation}
library(tibble)
library(ggplot2)
library(forcats)
dist = phyloseq::distance(ps_comp, method="euclidean")
ordination = ordinate(ps_comp, method="PCoA", distance=dist)
plot_ordination(ps_comp, ordination, color="Group") + 
    theme_classic() +
    theme(strip.background = element_blank()) + stat_ellipse(aes(group = Group), linetype = 2)

plot_ordination(ps_comp, ordination, color="Batch") + 
    theme_classic() +
    theme(strip.background = element_blank()) + stat_ellipse(aes(group = Batch), linetype = 2)
```

###PCoA clr: Group and batch
```{r PCoA representation}
dist = phyloseq::distance(ps_clr, method="euclidean")
ordination = ordinate(ps_clr, method="PCoA", distance=dist)
plot_ordination(ps_clr, ordination, color="Group") + 
    theme_classic() +
    theme(strip.background = element_blank()) + stat_ellipse(aes(group = Group), linetype = 2)

plot_ordination(ps_clr, ordination, color="Batch") + 
    theme_classic() +
    theme(strip.background = element_blank()) + stat_ellipse(aes(group = Batch), linetype = 2)
```

#Limma package- function: removebatcheffect
```{r Batch Correction}
library(limma)
library(mixOmics)
rawDada2_ASVs<-as.data.frame(otu_table(ps_clr))
y <- removeBatchEffect(((rawDada2_ASVs)), batch = metadata_hiper$Batch)
pca = pca(t(y), ncomp = 3, center = T, scale = F)
plotIndiv(pca, group = metadata_hiper$Group, legend = T, title = "removelima") 
plotIndiv(pca, group = metadata_hiper$Batch, legend = T, title = "removelima") 
```

##New Phyloseq object with transformed data(limma)
```{r new phyloseq} 
OTU_data_t <- as.data.frame((y))
#NewTax from ps_0. Needs to be run first
NewTAX<-tax_table(ps_filtered2)

OTU = otu_table(OTU_data_t, taxa_are_rows = TRUE)
TAX = NewTAX
SAMPLE <- sample_data(metadata_hiper)
phylo_tree <- readRDS("~/Documents/Lola/Datos_CIPF_2023/Rats_HE_HiperAmmon_model/outputs/phylo_tree.rds")
TREE = phylo_tree$tree

rownames(SAMPLE) <-metadata_hiper$SampleID
 
 # merge the data
 ps_limma <- phyloseq(OTU, TAX, SAMPLE, TREE)
 ps_limma
saveRDS(ps_limma,"~/Documents/Lola/Datos_CIPF_2023/Rats_HE_HiperAmmon_model/outputs/ps_limma.Rds")
```

###PCoA limma remove batch effect: Group and batch
```{r PCoA representation}
dist = phyloseq::distance(ps_limma, method="euclidean")
ordination = ordinate(ps_limma, method="PCoA", distance=dist)
plot_ordination(ps_limma, ordination, color="Group") + 
    theme_classic() +
    theme(strip.background = element_blank()) + stat_ellipse(aes(group = Group), linetype = 2)

plot_ordination(ps_limma, ordination, color="Batch") + 
    theme_classic() +
    theme(strip.background = element_blank()) + stat_ellipse(aes(group = Batch), linetype = 2)
```


###PCoA: centroid groups (limma remove batch effect)
```{r PCoA centroids}
##Euclidean distances
bray_matrix<- ps_limma@otu_table %>%
    t() %>%
    as.data.frame() %>%
    # euclidean dissimilarity
    vegan::vegdist(method = 'euclidean',
                   diag = TRUE, upper = TRUE)
metadata_hiper_h<-as.data.frame(sample_data(ps_limma))
metadata_hiper_h$Group[metadata_hiper_h$Group == "Hiperammonemic"] <- "Hyperammonemic"
rownames(metadata_hiper_h)<-metadata_hiper_h$SampleID
Group<-merge(as.data.frame(as.matrix(bray_matrix)), metadata_hiper_h, by="row.names")
Group<-as.vector(rownames(as.data.frame(as.matrix(bray_matrix))))

df <- metadata_hiper_h
df<-df[match(Group, df$SampleID),]

## Calculate multivariate dispersion or distance to the centroid
permdisp <- vegan::betadisper(bray_matrix, df$Group, "centroid")

##Plot
plot(permdisp, main="PCoA", hull=FALSE, ellipse=TRUE)
boxplot(permdisp, main="Distance to centroids")

axes<-permdisp$eig[1:3]/sum(permdisp$eig[permdisp$eig > 0]) * 100

plot(permdisp, main="PCoA clr", hull=FALSE, ellipse=TRUE, segments = TRUE, seg.col = c("#F8766D", "#00B0F6"), col = c("#F8766D", "#00B0F6"), label = TRUE, label.cex = 1.2, pch = c(20, 20, 20), seg.lty = c(1), seg.lwd = 1, cex.lab = 1.2, cex.main =1.3, xlab = paste0('PCoA1 (14.77%)'), ylab = paste0('PCoA2 (8.62 %)'))

permdisp$eig[1:3]/sum(permdisp$eig[permdisp$eig > 0]) * 100
#    PCoA1     PCoA2     PCoA3 
#14.767712  8.619812  7.391269 
```


###PCoA: centroid batch
```{r PCoA centroids}
##Euclidean distances
bray_matrix<- ps_limma@otu_table %>%
  t() %>%
  as.data.frame() %>%
  # euclidean dissimilarity
  vegan::vegdist(method = 'euclidean',
                 diag = TRUE, upper = TRUE)
metadata_hiper_h<-as.data.frame(sample_data(ps_limma))
rownames(metadata_hiper_h)<-metadata_hiper_h$SampleID
Batch<-merge(as.data.frame(as.matrix(bray_matrix)), metadata_hiper_h, by="row.names")
Batch<-as.vector(rownames(as.data.frame(as.matrix(bray_matrix))))

df <- metadata_hiper_h
df<-df[match(Batch, df$SampleID),]

## Calculate multivariate dispersion or distance to the centroid
permdisp <- vegan::betadisper(bray_matrix, df$Batch, "centroid")

##Plot
plot(permdisp, main="PCoA", hull=FALSE, ellipse=TRUE)
boxplot(permdisp, main="Distance to centroids")

axes<-permdisp$eig[1:3]/sum(permdisp$eig[permdisp$eig > 0]) * 100

plot(permdisp, main="PCoA relative abundance", hull=FALSE, ellipse=TRUE, segments = TRUE, seg.col = c("orange", "pink", "lightblue"), col = c("darkorange3", "magenta4", "mediumblue"), label = TRUE, label.cex = 1, pch = c(20, 20, 20), seg.lty = c(1), seg.lwd = 1, xlab = paste0('PCoA ',axes[1], '%'), ylab = paste0('PCoA ',axes[2], '%'))

permdisp$eig[1:3]/sum(permdisp$eig[permdisp$eig > 0]) * 100
```


#PERMANOVA 
```{r permanova types}
library(vegan)
#PERMANOVA
otu <- abundances(ps_limma)
meta <- meta(ps_limma)
permanova <- adonis(t(otu) ~ Group + Batch,
                    data = meta, permutations=999, method = "euclidean", by = "margin")

# P-value
permanova$aov.tab

#PERMANOVA having the Batch factor included 
otu <- data.frame(otu_table(ps_clr))
meta <- data.frame(sample_data(ps_clr))
permanova <- adonis(t(otu) ~ Group, strata = meta$Batch,
                    data = meta, permutations=999, method = "euclidean", by = "margin")

# P-value
permanova$aov.tab

# Analysis of Variance Table: homogeneidad de varianzas 
otu <- abundances(ps_limma)
dist <- vegdist(t(otu), method = "euclidean")
meta <- meta(ps_limma)
anova(betadisper(dist, meta$Group))

coef <- coefficients(permanova)["Group1",]
top.coef <- coef[rev(order(abs(coef)))[1:20]]
barplot(sort(top.coef), horiz = T, las = 1, main = "Top taxa")
```

#Comparison distance to centroids
```{r permanova Groups}
library(vegan)
library(usedist)
euclidean_matrix<- ps_limma@otu_table %>%
    t() %>%
    as.data.frame() %>%
    # euclidean distances
    vegan::vegdist(method = 'euclidean',
                   diag = TRUE, upper = TRUE)
metadata_hiper<-as.data.frame(sample_data(ps_limma))
rownames(metadata_hiper)<-metadata_hiper_h$SampleID
Group<-merge(as.data.frame(as.matrix(euclidean_matrix)), metadata_hiper, by="row.names")
Group<-as.vector(rownames(as.data.frame(as.matrix(euclidean_matrix))))

df <- metadata_hiper
df<-df[match(Group, df$SampleID),]

## Calculate multivariate dispersion or distance to the centroid
permdisp <- vegan::betadisper(euclidean_matrix, df$Group, "centroid")

TukeyHSD(permdisp)
anova(permdisp)

Hiperammonemic<-subset(df, Group=="Hiperammonemic")
Hiperammonemic<-as.vector(Hiperammonemic$SampleID)

Control<-subset(df, Group=="Control")
Control<-as.vector(Control$SampleID)

#Distance between groups of interest centroids
dist_between_centroids(euclidean_matrix, idx1=Control, idx2 = Hiperammonemic)

#Mean distances of samples to centroid of each group
permdisp[["group.distances"]]
usedist::dist_to_centroids(dist, df$Group)

```
