==> R/9_Network_analysis.Rmd <==
---
title: "9_Network_analysis"
author: "Lola Giner"
date: "2024-02-27"
output: html_document
---

devtools::install_github("stefpeschel/NetCoMi", 
                         dependencies = c("Depends", "Imports", "LinkingTo"),
                         repos = c("https://cloud.r-project.org/",
                                   BiocManager::repositories()))

devtools::install_github("zdk123/SpiecEasi")
devtools::install_github("GraceYoon/SPRING")

At OTU level you cannot see much, there are too many interactions
#Load packages
```{r Load packages}
library(NetCoMi)
library(phyloseq)
library(tidyr)
library(dplyr)
library(tibble)
library(purrr)
require(phyloseq)
require(tidyverse)
require(magrittr)
library(fastDummies)


get_latest_annotation <- function(phyloseq_obj) {
        tax_table <- phyloseq_obj@tax_table %>%
                as.data.frame() %>%
                rownames_to_column('ASV') %>%
                as_tibble() %>%
                tidyr::gather('Rank', 'Value', -ASV) %>%
                nest(data = -ASV) %>%
                mutate(TaxaID = map_chr(data, function(x) {
                        x %>%
                                dplyr::filter(!is.na(Value)) %>%
                                magrittr::use_series(Value) %>%
                                tail(1)
                })) %>%
                mutate(TaxaUp = map_chr(data, function(x) {
                        x %>%
                                dplyr::filter(!is.na(Value)) %>%
                                magrittr::use_series(Value) %>%
                                tail(2) %>% head(1)
                })) %>%
                mutate(TaxaID = paste(ASV, TaxaID)) %>%
                dplyr::select(-c(data, TaxaUp))
        
        return(tax_table)
}

```



#All genus network Pearson and ttest
We agglomerate by "genus"
```{r Pearson all}
ps_comp <- readRDS("~/Lola/Rats_HE_HiperAmmon_model/outputs/ps_comp.Rds")
metadata <- data.frame(sample_data(ps_comp))
metadata <- fastDummies::dummy_cols(metadata, select_columns = c("Group", "Batch")) 

# Keep columnsof interest
metadata_subset <- metadata[, c("Group_Control","Group_Hiperammonemic","Batch_B1","Batch_B2","Batch_B3")]


ps_genus <- tax_glom(ps_comp, taxrank = "Genus")
otu<-data.frame((otu_table(ps_genus)))
latest_annotations <- get_latest_annotation(ps_genus)
latest_annotations<-as.data.frame(latest_annotations)
rownames(otu)<-latest_annotations$TaxaID
otu_ok <- t(otu) 

data_input <- cbind(otu_ok, metadata_subset)
net_all <- netConstruct(data_input, measure = "pearson",
normMethod = "clr",
zeroMethod = "multRepl",
sparsMethod = "t-test",
verbose = 3,
dissFunc="unsigned",
seed = 123456)

props_all <- netAnalyze(net_all, clustMethod = "cluster_fast_greedy")

#Plot 
# Save the plot and the legend to an .svg file
svg("~/Lola/Rats_HE_HiperAmmon_model/outputs/Figures/Network_pearson_genus_ttest.svg", width = 10, height = 8)  # Specify the file name and dimensions
par(mfrow = c(1, 1))  # Set the layout to a single plot
plot<-plot(props_all, 
     nodeColor = "cluster", 
     nodeSize = "eigenvector",
     repulsion = 0.8,
     rmSingles = TRUE,
     labelScale = FALSE,
     cexLabels = 0.8,
     nodeSizeSpread = 5,
     cexNodes = 2,
     hubBorderCol = "black",
     posCol = "darkorange", 
     negCol = "darkturquoise",
     title1 = "Genus level Network with Pearson correlations", 
     showTitle = TRUE,
     cexTitle = 1.3)


legend(0.7, 1.0, cex = 0.80, title = "estimated correlation:",
       legend = c("+","-"), lty = 1.5, lwd = 2.5, col = c("darkorange","darkturquoise"),
       bty = "n", horiz = FALSE, x.intersp = 0.2, y.intersp = 0.5)
dev.off()  # Close the .svg device
#Save the plot to svg
```

#All family network
We agglomerate by "family"
```{r Pearson all}
ps_comp <- readRDS("~/Lola/Rats_HE_HiperAmmon_model/outputs/ps_comp.Rds")
metadata <- data.frame(sample_data(ps_comp))
metadata <- fastDummies::dummy_cols(metadata, select_columns = c("Group", "Batch")) 

# Keep columnsof interest
metadata_subset <- metadata[, c("Group_Control","Group_Hiperammonemic","Batch_B1","Batch_B2","Batch_B3")]


ps_genus <- tax_glom(ps_comp, taxrank = "Family")
otu<-data.frame((otu_table(ps_genus)))
latest_annotations <- get_latest_annotation(ps_genus)
latest_annotations<-as.data.frame(latest_annotations)
rownames(otu)<-latest_annotations$TaxaID
otu_ok <- t(otu) 

data_input <- cbind(otu_ok, metadata_subset)
net_all <- netConstruct(data_input, measure = "pearson",
normMethod = "clr",
zeroMethod = "multRepl",
sparsMethod = "t-test",
verbose = 3,
dissFunc="unsigned",
seed = 123456)

props_all <- netAnalyze(net_all, clustMethod = "cluster_fast_greedy")

#Plot 
# Save the plot and the legend to an .svg file
svg("~/Lola/Rats_HE_HiperAmmon_model/outputs/Figures/Network_pearson_family_ttest.svg", width = 10, height = 8)  # Specify the file name and dimensions
par(mfrow = c(1, 1))  # Set the layout to a single plot
plot<-plot(props_all, 
     nodeColor = "cluster", 
     nodeSize = "eigenvector",
     repulsion = 0.8,
     rmSingles = TRUE,
     labelScale = FALSE,
     cexLabels = 0.8,
     nodeSizeSpread = 5,
     cexNodes = 2,
     hubBorderCol = "black",
     posCol = "darkorange", 
     negCol = "darkturquoise",
     title1 = "Family level Network with Pearson correlations", 
     showTitle = TRUE,
     cexTitle = 1.3)


legend(0.7, 1.0, cex = 0.80, title = "estimated correlation:",
       legend = c("+","-"), lty = 1.5, lwd = 2.5, col = c("darkorange","darkturquoise"),
       bty = "n", horiz = FALSE, x.intersp = 0.2, y.intersp = 0.5)
dev.off()  # Close the .svg device
#Save the plot to svg
```

#Pearson Controls genus network
```{r Pearson Controls by genus network}
ps_clr <- readRDS("~/Lola/Rats_HE_HiperAmmon_model/outputs/ps_clr.Rds")
ps_genus <- tax_glom(ps_clr, taxrank = "Genus")
ps_genus_Controls <- subset_samples(ps_genus, Group == "Control")
otu<-data.frame((otu_table(ps_genus_Controls)))
latest_annotations <- get_latest_annotation(ps_genus_Controls)
latest_annotations<-as.data.frame(latest_annotations)
rownames(otu)<-latest_annotations$TaxaID
otu_ok <- t(otu) 

data_input <- cbind(otu_ok)

net_Controls <- netConstruct(data_input, 
              measure = "pearson", 
              taxRank = "Genus",
              normMethod = "none",
              zeroMethod = "none",
              sparsMethod = "t-test",
              verbose = 3,
              dissFunc="unsigned",
              seed = 123456)

props_Controls <- netAnalyze(net_Controls, clustMethod = "cluster_fast_greedy")
# Save the plot and the legend to an .svg file
svg("~/Lola/Rats_HE_HiperAmmon_model/outputs/Figures/Network_Controls.svg", width = 10, height = 8)  # Specify the file name and dimensions
par(mfrow = c(1, 1))  # Set the layout to a single plot
plot<-plot(props_Controls, 
     nodeColor = "cluster", 
     nodeSize = "eigenvector",
     repulsion = 0.8,
     rmSingles = TRUE,
     labelScale = FALSE,
     cexLabels = 0.8,
     nodeSizeSpread = 5,
     cexNodes = 2,
     hubBorderCol = "black",
     posCol = "darkturquoise", 
     negCol = "darkorange",
     title1 = "Controls:
     Genus level Network with Pearson correlations", 
     showTitle = TRUE,
     cexTitle = 1.3)


legend(0.7, 1.0, cex = 0.80, title = "estimated correlation:",
       legend = c("+","-"), lty = 1.5, lwd = 2.5, col = c("darkturquoise","darkorange"),
       bty = "n", horiz = FALSE, x.intersp = 0.2, y.intersp = 0.5)
dev.off()  # Close the .svg device
#Save the plot to svg
```

#PearsonHyperamonemic genus network
```{r Pearson NMCI by genus network}
ps_clr <- readRDS("~/Lola/Rats_HE_HiperAmmon_model/outputs/ps_clr.Rds")
ps_genus <- tax_glom(ps_clr, taxrank = "Genus")
ps_genus_Hyperamm <- subset_samples(ps_genus, Group == "Hiperammonemic")
otu<-data.frame((otu_table(ps_genus_Hyperamm)))
latest_annotations <- get_latest_annotation(ps_genus_Hyperamm)
latest_annotations<-as.data.frame(latest_annotations)
rownames(otu)<-latest_annotations$TaxaID
otu_ok <- t(otu) 

data_input <- cbind(otu_ok)

net_Hyperamm <- netConstruct(data_input, 
              measure = "pearson", 
              taxRank = "Genus",
              normMethod = "none",
              zeroMethod = "none",
              sparsMethod = "t-test",
              verbose = 3,
              dissFunc="unsigned",
              seed = 123456)

props_Hyperamm <- netAnalyze(net_Hyperamm, clustMethod = "cluster_fast_greedy")
# Save the plot and the legend to an .svg file
svg("~/Lola/Rats_HE_HiperAmmon_model/outputs/Figures/Network_Hyperammonemic.svg", width = 10, height = 8)  # Specify the file name and dimensions
par(mfrow = c(1, 1))  # Set the layout to a single plot
plot<-plot(props_Hyperamm, 
     nodeColor = "cluster", 
     nodeSize = "eigenvector",
     repulsion = 0.8,
     rmSingles = TRUE,
     labelScale = FALSE,
     cexLabels = 0.8,
     nodeSizeSpread = 5,
     cexNodes = 2,
     hubBorderCol = "black",
     posCol = "darkturquoise", 
     negCol = "darkorange",
     title1 = "Hyperammonemic:
     Genus level Network with Pearson correlations", 
     showTitle = TRUE,
     cexTitle = 1.3)


legend(0.7, 1.0, cex = 0.80, title = "estimated correlation:",
       legend = c("+","-"), lty = 1.5, lwd = 2.5, col = c("darkturquoise","darkorange"),
       bty = "n", horiz = FALSE, x.intersp = 0.2, y.intersp = 0.5)
dev.off()  # Close the .svg device
#Save the plot to svg
```


#SparCC Controls genus network
```{r SparCC Controls by genus network}
ps_filtered2 <- readRDS("~/Lola/Rats_HE_HiperAmmon_model/outputs/ps_filtered2.Rds")
ps_genus <- tax_glom(ps_filtered2, taxrank = "Genus")
ps_genus_Controls <- subset_samples(ps_genus, Group == "Control")
otu<-data.frame((otu_table(ps_genus_Controls)))
latest_annotations <- get_latest_annotation(ps_genus_Controls)
latest_annotations<-as.data.frame(latest_annotations)
rownames(otu)<-latest_annotations$TaxaID
rownames(otu) <- gsub("^ASV\\d+\\s", "", rownames(otu))

otu_ok <- t(otu) 

data_input <- cbind(otu_ok)

net_Controls <- netConstruct(data_input, 
                             measure = "sparcc", 
                             taxRank = "Genus",
                             normMethod = "clr",
                             zeroMethod = "none",
                             sparsMethod = "t-test",
                             verbose = 3,
                             dissFunc="unsigned",
                             seed = 123456)

props_Controls <- netAnalyze(net_Controls, clustMethod = "cluster_fast_greedy")
# Save the plot and the legend to an .svg file
svg("~/Lola/Rats_HE_HiperAmmon_model/outputs/Figures/Network_Controls_sparcc.svg", width = 10, height = 8)  # Specify the file name and dimensions
par(mfrow = c(1, 1))  # Set the layout to a single plot
plot<-plot(props_Controls, 
           nodeColor = "cluster", 
           nodeSize = "eigenvector",
           repulsion = 0.8,
           rmSingles = TRUE,
           labelScale = FALSE,
           cexLabels = 1.3,
           nodeSizeSpread = 5,
           cexNodes = 2,
           hubBorderCol = "black",
           posCol = "darkturquoise", 
           negCol = "darkorange",
           title1 = "Controls:
     Genus level Network with SparCC correlations", 
           showTitle = TRUE,
           cexTitle = 1.3)


legend(0.7, 1.0, cex = 0.80, title = "estimated correlation:",
       legend = c("+","-"), lty = 1.5, lwd = 2.5, col = c("darkturquoise","darkorange"),
       bty = "n", horiz = FALSE, x.intersp = 0.2, y.intersp = 0.5)
dev.off()  # Close the .svg device
#Save the plot to svg
```

#SparCCHyperamonemic genus network
```{r SparCC NMCI by genus network}
ps_filtered2 <- readRDS("~/Lola/Rats_HE_HiperAmmon_model/outputs/ps_filtered2.Rds")
ps_genus <- tax_glom(ps_filtered2, taxrank = "Genus")
ps_genus_Hyperamm <- subset_samples(ps_genus, Group == "Hiperammonemic")
otu<-data.frame((otu_table(ps_genus_Hyperamm)))
latest_annotations <- get_latest_annotation(ps_genus_Hyperamm)
latest_annotations<-as.data.frame(latest_annotations)
rownames(otu)<-latest_annotations$TaxaID
rownames(otu) <- gsub("^ASV\\d+\\s", "", rownames(otu))
otu_ok <- t(otu) 

data_input <- cbind(otu_ok)

net_Hyperamm <- netConstruct(data_input, 
                             measure = "sparcc", 
                             taxRank = "Genus",
                             normMethod = "clr",
                             zeroMethod = "none",
                             sparsMethod = "t-test",
                             verbose = 3,
                             dissFunc="unsigned",
                             seed = 123456)

props_Hyperamm <- netAnalyze(net_Hyperamm, clustMethod = "cluster_fast_greedy")
# Save the plot and the legend to an .svg file
svg("~/Lola/Rats_HE_HiperAmmon_model/outputs/Figures/Network_Hyperammonemic_sparcc.svg", width = 10, height = 8)  # Specify the file name and dimensions
par(mfrow = c(1, 1))  # Set the layout to a single plot
plot<-plot(props_Hyperamm, 
           nodeColor = "cluster", 
           nodeSize = "eigenvector",
           repulsion = 0.8,
           rmSingles = TRUE,
           labelScale = FALSE,
           cexLabels = 1.3,
           nodeSizeSpread = 5,
           cexNodes = 2,
           hubBorderCol = "black",
           posCol = "darkturquoise", 
           negCol = "darkorange",
           title1 = "Hyperammonemic:
     Genus level Network with SparCC correlations", 
           showTitle = TRUE,
           cexTitle = 1.3)


legend(0.7, 1.0, cex = 0.80, title = "estimated correlation:",
       legend = c("+","-"), lty = 1.5, lwd = 2.5, col = c("darkturquoise","darkorange"),
       bty = "n", horiz = FALSE, x.intersp = 0.2, y.intersp = 0.5)
dev.off()  # Close the .svg device
#Save the plot to svg
```
